requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langgraph
python-dotenv
pydantic
typing-extensions
httpx
google-generativeai
langchain_google_genai
langchain_tavily
